Added create writer
