+++ ASPCoreApiDemo.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Authors", "App\Authors.csproj", "{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{232A35F5-9712-4421-8D68-7747F24187B6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Debug|x64.Build.0 = Debug|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Debug|x86.Build.0 = Debug|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Release|x64.ActiveCfg = Release|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Release|x64.Build.0 = Release|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Release|x86.ActiveCfg = Release|Any CPU
+		{A87B1FBE-3BA7-4A96-BC32-FB7523DD4778}.Release|x86.Build.0 = Release|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Debug|x64.Build.0 = Debug|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Debug|x86.Build.0 = Debug|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Release|x64.ActiveCfg = Release|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Release|x64.Build.0 = Release|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Release|x86.ActiveCfg = Release|Any CPU
+		{232A35F5-9712-4421-8D68-7747F24187B6}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
