+++ App/Controllers/AuthorsController.cs
@@ -1,5 +1,7 @@
 using System.Linq;
 using Microsoft.AspNetCore.Mvc;
+using Authors.Repository;
+using Authors.Models;
 
 namespace Authors.Controllers
 {
@@ -56,6 +58,18 @@ namespace Authors.Controllers
             }
 
              return Ok(writer);
+        }
+
+        [HttpPost("/Author")]
+        [ProducesResponseType(202)]
+        public IActionResult Insert([FromBody] Writer author)
+        {
+            if (this.libraryRepository.AddWriter(author) == 1)
+            {
+                return this.Accepted();
+            }
+
+            return BadRequest();
         }
     }
 }
