+++ App/Repository/ILibraryRepository.cs
@@ -1,11 +1,12 @@
 using System.Collections.Generic;
 using Authors.Models;
 
-namespace Authors.Controllers
+namespace Authors.Repository
 {
     public interface ILibraryRepository
     {
         IEnumerable<Writer> GetWriters();
         Writer GetWriter(long id);
+        int AddWriter(Writer author);
     }
 }
