+++ App/Repository/LibraryRepository.cs
@@ -1,3 +1,4 @@
+﻿using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
@@ -5,31 +6,8 @@ using Authors.Models;
 using Dapper;
 using Microsoft.Extensions.Configuration;
 
-namespace Authors.Controllers
+namespace Authors.Repository
 {
-    public class LibraryRepositoryMock : ILibraryRepository
-    {
-        public IEnumerable<Writer> GetWriters()
-        {
-            return new List<Writer> {
-                new Writer("Stephen King", null),
-                new Writer("Neil Gaiman", null)
-            };
-        }
-
-        public Writer GetWriter(long id)
-        {
-            if (id == 0)
-            {
-                return null;
-            }
-
-            return new Writer {
-                Id = id
-            };
-        }
-    }
-
     public class LibraryRepository : ILibraryRepository
     {
         private readonly IConfiguration config;
@@ -49,7 +27,7 @@ namespace Authors.Controllers
 
         public IEnumerable<Writer> GetWriters()
         {
-            using (IDbConnection db = Connection)
+            using (var db = Connection)
             {
                 return db.Query<Writer>("Select * From Writer");
             }
@@ -57,11 +35,25 @@ namespace Authors.Controllers
 
         public Writer GetWriter(long id)
         {
-            using (IDbConnection db = Connection)
+            using (var db = Connection)
             {
-                var writer = db.QueryFirstOrDefault<Writer>("Select * From Writer where id =@Id", new {Id = id});
+                var writer = db.QueryFirstOrDefault<Writer>("Select * From Writer where id =@Id", new { Id = id });
                 return writer;
             }
         }
+
+        public int AddWriter(Writer writer)
+        {
+            if (writer == null)
+            {
+                throw new ArgumentNullException(nameof(writer));
+            }
+
+            using (var db = Connection)
+            {
+                var result = db.Execute("Insert Into Writer (Name, DateOfBirth) values (@name, @dateOfBirth)", new { Name = writer.Name, DateOfBirth = writer.DateOfBirth });
+                return result;
+            }
+        } 
     }
 }
